feat: add lazy loading for rest-client page
