--- src/app/[locale]/(privat)/variables/components/VariableRow.tsx
@@ -1,12 +1,0 @@
-'use client';
-import { useTranslations } from 'next-intl';
-
-export default function VariablesList() {
-  const t = useTranslations('vars');
-
-  return (
-    <>
-      <h1>{t('title')}</h1>
-    </>
-  );
-}
